Changes: 	Separated out the Settings window into multiple class still buggy Issues: 	Found you that you cant kill a QRunnable (WWWWWWHHHHHYYYYY!!!!!!)
